+++ src/app/components/CarouselWithContent.tsx
