+++ src/app/components/EmailTemplate.tsx
@@ -1,4 +1,3 @@
-"use client";
 import * as React from "react";
 
 interface EmailTemplateProps {
