+++ src/app/utils/shared.ts
