--- src/pages/_app.tsx
@@ -1,17 +1,0 @@
-import "@/styles/globals.css";
-import { useEffect } from "react";
-import AOS from "aos";
-import "aos/dist/aos.css";
-import type { AppProps } from "next/app";
-import { ThemeProvider } from "@material-tailwind/react";
-
-export default function App({ Component, pageProps }: AppProps) {
-  useEffect(() => {
-    AOS.init();
-  }, []);
-  return (
-    <ThemeProvider>
-      <Component {...pageProps} />
-    </ThemeProvider>
-  );
-}
